+++ Cesium3DTilesSelection/src/RasterizedPolygonsOverlay.cpp
@@ -22,9 +22,21 @@ void rasterizePolygons(
     LoadedRasterOverlayImage& loaded,
     const CesiumGeospatial::GlobeRectangle& rectangle,
     const glm::dvec2& textureSize,
-    const std::vector<CartographicPolygon>& cartographicPolygons) {
+    const std::vector<CartographicPolygon>& cartographicPolygons,
+    bool invertSelection) {
 
   CesiumGltf::ImageCesium& image = loaded.image.emplace();
+
+  std::byte insideColor;
+  std::byte outsideColor;
+
+  if (invertSelection) {
+    insideColor = static_cast<std::byte>(0);
+    outsideColor = static_cast<std::byte>(0xff);
+  } else {
+    insideColor = static_cast<std::byte>(0xff);
+    outsideColor = static_cast<std::byte>(0);
+  }
 
   // create a 1x1 mask if the rectangle is completely inside a polygon
   if (Cesium3DTilesSelection::CesiumImpl::withinPolygons(
@@ -35,10 +47,7 @@ void rasterizePolygons(
     image.height = 1;
     image.channels = 1;
     image.bytesPerChannel = 1;
-    image.pixelData.resize(1);
-
-    image.pixelData[0] = static_cast<std::byte>(0xff);
-
+    image.pixelData.resize(1, insideColor);
     return;
   }
 
@@ -61,8 +70,7 @@ void rasterizePolygons(
     image.height = 1;
     image.channels = 1;
     image.bytesPerChannel = 1;
-    image.pixelData.resize(1);
-
+    image.pixelData.resize(1, outsideColor);
     return;
   }
 
@@ -75,7 +83,7 @@ void rasterizePolygons(
   image.height = int32_t(glm::round(textureSize.y));
   image.channels = 1;
   image.bytesPerChannel = 1;
-  image.pixelData.resize(size_t(image.width * image.height));
+  image.pixelData.resize(size_t(image.width * image.height), outsideColor);
 
   // TODO: this is naive approach, use line-triangle
   // intersections to rasterize one row at a time
@@ -138,40 +146,13 @@ void rasterizePolygons(
                v_proj_bc_perp >= 0.0) ||
               (v_proj_ab_perp <= 0.0 && v_proj_ca_perp <= 0.0 &&
                v_proj_bc_perp <= 0.0)) {
-            image.pixelData[width * j + i] = static_cast<std::byte>(0xff);
+            image.pixelData[width * j + i] = insideColor;
           }
         }
       }
     }
   }
 }
-
-Rectangle computeCoverageRectangle(
-    const Projection& projection,
-    const std::vector<CartographicPolygon>& polygons) {
-  std::optional<GlobeRectangle> result;
-
-  for (const CartographicPolygon& polygon : polygons) {
-    std::optional<GlobeRectangle> maybeRectangle =
-        polygon.getBoundingRectangle();
-    if (!maybeRectangle) {
-      continue;
-    }
-
-    if (result) {
-      result = result->computeUnion(*maybeRectangle);
-    } else {
-      result = maybeRectangle;
-    }
-  }
-
-  if (result) {
-    return projectRectangleSimple(projection, *result);
-  } else {
-    return Rectangle(0.0, 0.0, 0.0, 0.0);
-  }
-}
-
 } // namespace
 
 class CESIUM3DTILESSELECTION_API RasterizedPolygonsTileProvider final
@@ -179,6 +160,7 @@ class CESIUM3DTILESSELECTION_API RasterizedPolygonsTileProvider final
 
 private:
   std::vector<CartographicPolygon> _polygons;
+  bool _invertSelection;
 
 public:
   RasterizedPolygonsTileProvider(
@@ -189,7 +171,8 @@ public:
           pPrepareRendererResources,
       const std::shared_ptr<spdlog::logger>& pLogger,
       const CesiumGeospatial::Projection& projection,
-      const std::vector<CartographicPolygon>& polygons)
+      const std::vector<CartographicPolygon>& polygons,
+      bool invertSelection)
       : RasterOverlayTileProvider(
             owner,
             asyncSystem,
@@ -198,8 +181,16 @@ public:
             pPrepareRendererResources,
             pLogger,
             projection,
-            computeCoverageRectangle(projection, polygons)),
-        _polygons(polygons) {}
+            // computeCoverageRectangle(projection, polygons)),
+            projectRectangleSimple(
+                projection,
+                CesiumGeospatial::GlobeRectangle(
+                    -CesiumUtility::Math::OnePi,
+                    -CesiumUtility::Math::PiOverTwo,
+                    CesiumUtility::Math::OnePi,
+                    CesiumUtility::Math::PiOverTwo))),
+        _polygons(polygons),
+        _invertSelection(invertSelection) {}
 
   virtual CesiumAsync::Future<LoadedRasterOverlayImage>
   loadTileImage(RasterOverlayTile& overlayTile) override {
@@ -212,6 +203,7 @@ public:
 
     return this->getAsyncSystem().runInWorkerThread(
         [&polygons = this->_polygons,
+         invertSelection = this->_invertSelection,
          projection = this->getProjection(),
          rectangle = overlayTile.getRectangle(),
          textureSize]() -> LoadedRasterOverlayImage {
@@ -221,7 +213,12 @@ public:
           LoadedRasterOverlayImage result;
           result.rectangle = rectangle;
 
-          rasterizePolygons(result, tileRectangle, textureSize, polygons);
+          rasterizePolygons(
+              result,
+              tileRectangle,
+              textureSize,
+              polygons,
+              invertSelection);
 
           return result;
         });
@@ -231,11 +228,13 @@ public:
 RasterizedPolygonsOverlay::RasterizedPolygonsOverlay(
     const std::string& name,
     const std::vector<CartographicPolygon>& polygons,
+    bool invertSelection,
     const CesiumGeospatial::Ellipsoid& ellipsoid,
     const CesiumGeospatial::Projection& projection,
     const RasterOverlayOptions& overlayOptions)
     : RasterOverlay(name, overlayOptions),
       _polygons(polygons),
+      _invertSelection(invertSelection),
       _ellipsoid(ellipsoid),
       _projection(projection) {}
 
@@ -261,7 +260,8 @@ RasterizedPolygonsOverlay::createTileProvider(
               pPrepareRendererResources,
               pLogger,
               this->_projection,
-              this->_polygons));
+              this->_polygons,
+              this->_invertSelection));
 }
 
 } // namespace Cesium3DTilesSelection